remove reference lister from checkers
A violation of a file is a violation regardless of whether the violation
has been previously recorded in the deprecated_references.yml
file. This change removes the dependency on ReferenceLister
for the checkers. While not strictly necessary for making
Packwerk parallel, it greatly simplifies our reasoning as
checking for existing deprecations can be done on the offense
list as a whole rather than when we decide to make an offense or not.