update-deprecations should always return true
currently in production we always return true when running
update-deprecations because no violation fails the check
in `CacheDeprecatedReferences#listed?`. We only test for
this command to return false in one contrived test that
never actually occurs so it has been removed.

All messaging has been updated to remove the string
indicating no offenses detected as that makes no sense
in this context.