introduce OffenseCollection
finding the offenses for a ParseRun now returns an OffenseCollection object.
This refactor is made possible by amalgamating the CacheDeprecatedReferences
class and all its subclasses into a single OffenseCollection class.

OffenseCollections know about the different between errors and reference
violations which is something that was handled implicitly before. We now
no longer need to mutate the new entries in a method called `listed?`
just to satisfy an interface.

This change makes it much more clear what each method is really doing
and simplifies the code for future readers.